"working" render of fallingLetters component.
