+++ packages/courseware/src/typing/questions/falling-letters/FallingLetters.vue
@@ -61,7 +61,7 @@
 <script lang="ts">
 import { defineComponent, ref, onMounted, onUnmounted, PropType } from 'vue';
 import { useViewable, useQuestionView } from '@vue-skuilder/common-ui';
-import { FallingLettersQuestion } from './index';
+import { FallingLettersQuestion } from './FallingLettersQuestion';
 import { ViewData } from '@vue-skuilder/common';
 import { Letter, TreePosition } from './types';
